Update to 2.6.5 and make build work on el7 and el6
This required jglobus 2.1.0, so added that version as an explicit
requirement.  The maven artifact names for emi-trustmanager and
voms-api-java had changed and needed to be updated.

The patches to remove the sunxacml requirement and update to
bouncycastle 1.47+ were removed, since they were no longer necessary:
privilege-xacml-2.6.5 no longer uses bouncycastle at all.  The
requirement has been removed from the specfile and a patch was added to
remove it from the pom.xml.

The Require/BuildRequire for slf4j was removed and instead it's brought
in by maven and bundled.  This was the easiest way of making the build
work for el6 and el7, and the version available on the system (1.5.8 for
el6 and 1.7.4 for el7) was not what was actually required by the pom.xml
file (1.5.10).

git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat@21680 4e558342-562e-0410-864c-e07659590f8c